tracefs: Use libtracefs and libtraceevent
It's been over 10 years that powertop used a copy of libtraceevent as a
temporary case until the libtraceevent library was ready for use. Well,
it's been ready for use for several years now. libtracefs has also been
added, which is a superset of libtraceevent and takes care of accessing
files in the tracefs file system (and will even find it for you).

Currently, this update only converts over to using the libraries and
removes its internal copy. It does not yet take advantage of libtracefs,
but will shortly.

Note, some naming convention updates have taken place in the creation of
libtracefs. Namely, "pevent" is no longer used, and instead "tep" is (for
trace event parser).

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>