tracefs: Use tracefs_event_file_read() to find the events
Use the libtracefs tracefs_event_file_read() function to find the format
file of the event. It simplifies the code as it handle not only allocating
the buffer to pass the file into, but also will search for where the
tracefs file system is mounted. As not all systems mount debugfs anymore
(tracefs is its own file system and some distros only mount it in
/sys/kernel/tracing).

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>